smartquail-dashboard/requirements.txt
# SmartQuail Dashboard Dependencies
# ==================================

# Core
streamlit>=1.28.0

# Database
supabase>=2.0.0

# MQTT
paho-mqtt>=1.6.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
